# packages/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CariCash Nova API
  version: 0.2.0
  description: |
    CariCash Nova — Barbados-based mobile payments platform API.

    ## Key Concepts
    - **Jurisdiction**: Barbados. Base currency: BBD. NO FX — cross-currency postings rejected.
    - **Closed-loop wallets**: Customers, agents, and merchants hold balances internally.
    - **Double-entry ledger**: Every transaction posts balanced journal entries (sum DR = sum CR).
    - **Idempotency**: All money-moving endpoints require `idempotency_key`. Retries with same key+payload return cached result; same key+different payload returns 409 DUPLICATE_IDEMPOTENCY_CONFLICT.
    - **Maker-checker**: Reversals, manual adjustments, overdraft facilities, fee/commission changes require dual approval (maker ≠ checker).
    - **Posting via Durable Object**: All money movement routes through a serialized Durable Object per wallet-per-currency.
  contact:
    name: CariCash Engineering
    email: engineering@caricash.bb
  license:
    name: MIT
servers:
  - url: /
    description: Current host
  - url: http://localhost:8787
    description: Local development (wrangler dev)
...
security: []
paths:
...
  /customers:
    post:
      operationId: createCustomer
      tags: [Customers]
      summary: Register a new customer
      description: |
        Create a customer account with name details, preferred display name, and registration metadata.
        Automatically creates a BBD wallet and registration metadata record.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
            example:
              msisdn: '+12465559999'
              name: 'Jane Marie Doe'
              first_name: 'Jane'
              middle_name: 'Marie'
              last_name: 'Doe'
              preferred_name: 'FIRST_NAME'
              display_name: 'Jane'
              registration_type: SELF_REGISTRATION
              channel: WEB
              terms_accepted: true
              privacy_accepted: true
              marketing_opt_in: false
              pin: '1234'
              correlation_id: '01HZ...'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCustomerResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Customer with this MSISDN already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /customers/{customerId}:
    get:
      operationId: getCustomer
      tags: [Customers]
      summary: Get customer details
      description: Retrieve customer profile by ID. Returns 501 (stub).
      parameters:
        - $ref: '#/components/parameters/customerId'
      responses:
        '200':
          description: Customer details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '501':
          description: Not yet implemented
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /customers/{customerId}/kyc/initiate:
    post:
      operationId: initiateCustomerKyc
      tags: [Customers]
      summary: Initiate customer KYC
      description: Submit KYC documents for a customer. Updates kyc_state to PENDING.
      parameters:
        - $ref: '#/components/parameters/customerId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KycInitiateRequest'
            example:
              document_type: 'NATIONAL_ID'
              document_number: 'BB123456'
              correlation_id: '01HZ...'
      responses:
        '200':
          description: KYC initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KycInitiateResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
  /customers/{customerId}/kyc:
    get:
      operationId: getCustomerKyc
      tags: [Customers]
      summary: Get customer KYC status
      description: Retrieve KYC status for a customer. Returns 501 (stub).
      parameters:
        - $ref: '#/components/parameters/customerId'
      responses:
        '200':
          description: KYC profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KycProfileResponse'
        '501':
          description: Not yet implemented
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================
  #  AGENTS
  # ============================================================
...
components:
  securitySchemes:
    CustomerAuth:
      type: http
      scheme: bearer
      description: Bearer token obtained from POST /auth/customer/login
    AgentAuth:
      type: http
      scheme: bearer
      description: Bearer token obtained from POST /auth/agent/login
    MerchantAuth:
      type: http
      scheme: bearer
      description: Bearer token obtained from POST /auth/merchant/login
    StaffAuth:
      type: http
      scheme: bearer
      description: Bearer token obtained from POST /auth/staff/login (OAuth stub)
  parameters:
    customerId:
      name: customerId
      in: path
      required: true
      schema:
        type: string
      description: Customer actor ID (ULID)
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'Validation failed'
            code: VALIDATION_ERROR
            correlation_id: '01HZ...'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: 'Not found'
            code: NOT_FOUND
            correlation_id: '01HZ...'
...
  schemas:
...
    ActorType:
      type: string
      enum: [CUSTOMER, AGENT, MERCHANT, STAFF]
...
    RegistrationType:
      type: string
      enum: [SELF_REGISTRATION, AGENT_REGISTRATION, STAFF_REGISTRATION, BULK_IMPORT, API_INTEGRATION, MERCHANT_REFERRAL]
    RegistrationChannel:
      type: string
      enum: [USSD, APP, WEB, API, PORTAL, IN_PERSON]
    PreferredNameSource:
      type: string
      enum: [FIRST_NAME, MIDDLE_NAME, LAST_NAME, FULL_NAME, CUSTOM]
...
    KycStatus:
      type: string
      enum: [NOT_STARTED, PENDING, APPROVED, REJECTED]
    ActorState:
      type: string
      enum: [PENDING, ACTIVE, SUSPENDED, CLOSED]
...
    ErrorCodeEnum:
      type: string
      enum:
        - UNAUTHORIZED
        - FORBIDDEN
        - VALIDATION_ERROR
        - INSUFFICIENT_FUNDS
        - CROSS_CURRENCY_NOT_ALLOWED
        - LIMIT_EXCEEDED
        - KYC_REQUIRED
        - BLOCKLISTED
        - DUPLICATE_IDEMPOTENCY_CONFLICT
        - IDEMPOTENCY_IN_PROGRESS
        - MAKER_CHECKER_REQUIRED
        - NOT_FOUND
        - CONFLICT
        - INTERNAL_ERROR
        - UNBALANCED_JOURNAL
        - JOURNAL_NOT_FOUND
        - ACCOUNT_NOT_FOUND
        - MAKER_CHECKER_VIOLATION
        - APPROVAL_NOT_FOUND
        - APPROVAL_ALREADY_DECIDED
        - INVALID_CREDENTIALS
        - ACCOUNT_LOCKED
        - RATE_LIMIT_EXCEEDED
        - SESSION_EXPIRED
        - ACTOR_NOT_FOUND
        - ACTOR_SUSPENDED
        - DUPLICATE_IDENTIFIER
        - MISSING_REQUIRED_FIELD
        - RECONCILIATION_MISMATCH
        - INTEGRITY_VIOLATION
        - REPAIR_FAILED

    # ---- Error response ----
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          $ref: '#/components/schemas/ErrorCodeEnum'
        name:
          type: string
          description: Error class name
        correlation_id:
          type: string
        details:
          type: object
          additionalProperties: true
        retryable:
          type: boolean
          default: false

    # ---- Auth requests ----
...
    Actor:
      type: object
      properties:
        id:
          type: string
        type:
          $ref: '#/components/schemas/ActorType'
        state:
          $ref: '#/components/schemas/ActorState'
        msisdn:
          type: string
        agent_code:
          type: string
        store_code:
          type: string
        staff_code:
          type: string
        name:
          type: string
        first_name:
          type: string
        middle_name:
          type: string
        last_name:
          type: string
        display_name:
          type: string
        email:
          type: string
          format: email
        kyc_state:
          $ref: '#/components/schemas/KycStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ActorResponse:
      type: object
      properties:
        actor:
          $ref: '#/components/schemas/Actor'
        correlation_id:
          type: string

    # ---- Customer ----
    CreateCustomerRequest:
      type: object
      required: [msisdn, name, pin]
      properties:
        msisdn:
          type: string
        name:
          type: string
        first_name:
          type: string
        middle_name:
          type: string
        last_name:
          type: string
        preferred_name:
          $ref: '#/components/schemas/PreferredNameSource'
        display_name:
          type: string
        email:
          type: string
          format: email
        pin:
          type: string
        registration_type:
          $ref: '#/components/schemas/RegistrationType'
        channel:
          $ref: '#/components/schemas/RegistrationChannel'
        registered_by_actor_id:
          type: string
        referral_code:
          type: string
        campaign_id:
          type: string
        terms_accepted:
          type: boolean
        privacy_accepted:
          type: boolean
        marketing_opt_in:
          type: boolean
        correlation_id:
          type: string
        idempotency_key:
          type: string
    CreateCustomerResponse:
      type: object
      properties:
        actor:
          $ref: '#/components/schemas/Actor'
        wallet_id:
          type: string
        registration_id:
          type: string
        registration_type:
          $ref: '#/components/schemas/RegistrationType'
        correlation_id:
          type: string

    # ---- Agent ----
...
    KycInitiateRequest:
      type: object
      required: [document_type, document_number]
      properties:
        document_type:
          type: string
          description: Document type (e.g., NATIONAL_ID, PASSPORT)
        document_number:
          type: string
        correlation_id:
          type: string
    KycInitiateResponse:
      type: object
      properties:
        actor_id:
          type: string
        kyc_state:
          $ref: '#/components/schemas/KycStatus'
        correlation_id:
          type: string
    KycProfileResponse:
      type: object
      properties:
        actor_id:
          type: string
        kyc_state:
          $ref: '#/components/schemas/KycStatus'
        document_type:
          type: string
        document_number:
          type: string

    # ---- Wallet ----
...